--- src/components/index.tsx
@@ -1,6 +1,0 @@
-export { Title } from './title'
-export { Form } from './form'
-export { Alert } from './alert'
-export { Download } from './download'
-export { Loader } from './loader'
-export { Controller } from './controller'
